+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
-* Add ability to clear the boolean gate only (useful for debugging)
+* Add ability to clear the boolean gate only (useful for debugging).
 * Added `FunWithFlags.get_flag/1`, to retrieve a flag struct. Useful for debugging.
+* Internal improvements.
 
 ## v0.10.1