+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v0.11.0
 
 * Add ability to clear the boolean gate only (useful for debugging).
 * Added `FunWithFlags.get_flag/1`, to retrieve a flag struct. Useful for debugging.